pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "artsy"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "mlops-28", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "dvc-gs>=3.0.2",
    "fastapi==0.115.6",
    "pytorch-lightning>=2.6.0",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "typer==0.21.1",
    "uvicorn==0.34.0",
    "pathspec<0.12",
    "hydra-core>=1.3.2",
    "utils>=1.0.2",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.14.13",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
